mako_hotplug: some nice improvements
- automatically online a cpu with a boosted frequency instead of waiting
a sample for recalculation of load bla bla. If a core goes online it should be
because there is a bigger load so to save ms and to get better performance
the core comes online boosted to hispeed_freq that comes from interactive gov;

- wait 5 seconds before scaling up interactive tunables on high load. Sometimes
just by interacting with the UI around we can *ocasionally* have the four cores
online, but we don't really need the high performance tunables. If we're gaming
we'll hit this 5 seconds easily;

- new first_level default, decreased from 90 to 80;

Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>